version 1.6.9: make yaml2db faster by skipping unchanged keys

+++ lib/tasks/translation_center.rake
@@ -41,15 +41,10 @@ namespace :translation_center do
   task :synch, [:locale ] => :environment do |t, args|
     begin
       if TranslationCenter::Category.any?
-        puts "WARNING: You already have translations stored in the db, this action will destroy them. Are you sure, you want to continue? press [Y|n]"
+        puts "WARNING: You already have translations stored in the db, do you want to destroy them? press [Y|n]"
         confirm = $stdin.gets.chomp
     
-        if confirm.blank? || confirm == 'y' || confirm == 'yes'
-          TranslationCenter::Category.destroy_all
-        else
-          next
-        end
-
+        TranslationCenter::Category.destroy_all if confirm.blank? || confirm == 'y' || confirm == 'yes'
       end
       TranslationCenter.yaml2db(args[:locale])
       TranslationCenter.db2yaml(args[:locale])
